requirements.txt
mysql-connector-python==8.2.0
protobuf==4.21.12
termcolor==2.4.0
